SLM-81 Rename newContactValidator
